scripts/requirements.txt
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
matplotlib==3.9.2